requirements.txt
asgiref==3.11.0
Django==5.2.8
sqlparse==0.5.3
gunicorn==23.0.0
djangorestframework==3.15.0
